Rename openstack-common to oslo.
At the Design Summit it was decided that the OpenStack common modules
would be named Oslo as part of an effort to convert them into a core
project housing a collection of library dependencies. This handles the
Jenkins job renames and git project name references within jobs.

Change-Id: I46ddcdd989dca85e061b23de1045c763ae2a24af
Reviewed-on: https://review.openstack.org/15836
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Approved: Jeremy Stanley <fungi@yuggoth.org>
Reviewed-by: Jeremy Stanley <fungi@yuggoth.org>
Tested-by: Jenkins