Fix defaults for gerritbot class parameters
Without this patch, the gerritbot module makes the nick, password,
server, vhost_name, user, and channel_file parameters optional by
providing default values for them. The default values are empty
strings, which are not valid and will cause the bot to not function
properly. This patch changes the vhost_name paremeter to have a default
value of $::fqdn, which is a reasonable default, and makes the other
parameters required by removing the defaults since there are no sane
defaults to provide.

This will mean that puppet runs will fail if no values are supplied to
these parameters. However, Infra is already providing values in
system-config, and other users of this module must be providing values
in order for gerritbot to work, so this should not be a destructive
change.

Change-Id: Ia0eed0f3e8341715236c7dee78840ed4443ddf90