Refactor bot into defined types to allow multiple bots
The Openstack-Infra project recently hit the limit of channels that the
gerritbot can be joined on Freenode. As discussed, the desired action
for here is to spin up a 'gerribot2'. However, due to how the
puppet-gerritbot module was constructed, this is not quite simple.

This commit refactors the 'bot' in this module to be a defined type,
supporting multiple init scripts, config files, and logging files. This
way, a second 'gerritbot2' can be spun up to join more channels.

Change-Id: I5a8cd48c13a9a750d0b5a301a3c7202c97ca9f22